+++ .env
@@ -1,3 +1,4 @@
+VUE_APP_ENV=production
 VUE_APP_NETWORK=homestead
 IPFS_DEPLOY_PINATA__API_KEY=
 IPFS_DEPLOY_PINATA__SECRET_API_KEY=
