+++ src/config/homestead.staging.json
@@ -1,0 +1,3 @@
+{
+  "subgraphUrl": "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-beta"
+}
