+++ src/config/kovan.json
@@ -1,0 +1,135 @@
+{
+  "network": "kovan",
+  "chainId": 42,
+  "defaultPrecision": 4,
+  "subgraphUrl": "https://api.thegraph.com/subgraphs/name/balancer-labs/balancer-kovan",
+  "alchemyWsUrl": "wss://eth-kovan.ws.alchemyapi.io/v2/C467LJc7Aa761LLxlcxXvhIG9HzDTtKw",
+  "addresses": {
+    "bFactory": "0x8f7F78080219d4066A8036ccD30D588B416a40DB",
+    "bActions": "0x1266F7220cCc4f1957b3f1EA713F3F434Fc3BDc7",
+    "dsProxyRegistry": "0x130767E0cf05469CF11Fa3fcf270dfC1f52b9072",
+    "proxy": "0xdf6F5B1f49eFf8d1C45f2AADC045cB91C33e44b8",
+    "weth": "0xd0A1E359811322d97991E03f863a0C30C2cF029C",
+    "multicall": "0x2cc8688C5f75E365aaEEb4ea8D6a480405A48D2A"
+  },
+  "connectors": {
+    "injected": {
+      "id": "injected",
+      "name": "MetaMask"
+    },
+    "walletconnect": {
+      "id": "walletconnect",
+      "name": "WalletConnect"
+    },
+    "portis": {
+      "id": "portis",
+      "name": "Portis",
+      "options": {
+        "network": "kovan",
+        "dappId": "3f1c3cfc-7dd5-4e8a-aa03-71ff7396d9fe"
+      }
+    },
+    "walletlink": {
+      "id": "walletlink",
+      "name": "Coinbase",
+      "options": {
+        "appName": "Balancer Pool Management",
+        "darkMode": true,
+        "chainId": 42,
+        "ethJsonrpcUrl": "https://eth-kovan.alchemyapi.io/v2/C467LJc7Aa761LLxlcxXvhIG9HzDTtKw"
+      }
+    }
+  },
+  "tokens": {
+    "0xd0A1E359811322d97991E03f863a0C30C2cF029C": {
+      "address": "0xd0A1E359811322d97991E03f863a0C30C2cF029C",
+      "symbol": "WETH",
+      "id": "weth",
+      "decimals": 18,
+      "precision": 4,
+      "chartColor": "#828384"
+    },
+    "0x1528F3FCc26d13F7079325Fb78D9442607781c8C": {
+      "address": "0x1528F3FCc26d13F7079325Fb78D9442607781c8C",
+      "symbol": "DAI",
+      "id": "dai",
+      "decimals": 18,
+      "precision": 2,
+      "chartColor": "#fab323"
+    },
+    "0xef13C0c8abcaf5767160018d268f9697aE4f5375": {
+      "address": "0xef13C0c8abcaf5767160018d268f9697aE4f5375",
+      "symbol": "MKR",
+      "id": "maker",
+      "decimals": 18,
+      "precision": 4,
+      "chartColor": "#5eb8aa"
+    },
+    "0x2F375e94FC336Cdec2Dc0cCB5277FE59CBf1cAe5": {
+      "address": "0x2F375e94FC336Cdec2Dc0cCB5277FE59CBf1cAe5",
+      "symbol": "USDC",
+      "id": "usd-coin",
+      "decimals": 6,
+      "precision": 2,
+      "chartColor": "#2875ca"
+    },
+    "0x8c9e6c40d3402480ACE624730524fACC5482798c": {
+      "address": "0x8c9e6c40d3402480ACE624730524fACC5482798c",
+      "symbol": "REP",
+      "id": "augur",
+      "decimals": 18,
+      "precision": 4,
+      "chartColor": "#422940"
+    },
+    "0xe0C9275E44Ea80eF17579d33c55136b7DA269aEb": {
+      "address": "0xe0C9275E44Ea80eF17579d33c55136b7DA269aEb",
+      "symbol": "WBTC",
+      "id": "wrapped-bitcoin",
+      "decimals": 8,
+      "precision": 4,
+      "chartColor": "#779bc0"
+    },
+    "0x1f1f156E0317167c11Aa412E3d1435ea29Dc3cCE": {
+      "address": "0x1f1f156E0317167c11Aa412E3d1435ea29Dc3cCE",
+      "symbol": "BAT",
+      "id": "basic-attention-token",
+      "decimals": 18,
+      "precision": 2,
+      "chartColor": "#ff5700"
+    },
+    "0x86436BcE20258a6DcfE48C9512d4d49A30C4d8c4": {
+      "address": "0x86436BcE20258a6DcfE48C9512d4d49A30C4d8c4",
+      "symbol": "SNX",
+      "id": "havven",
+      "decimals": 18,
+      "precision": 2,
+      "chartColor": "#ffffff"
+    },
+    "0x37f03a12241E9FD3658ad6777d289c3fb8512Bc9": {
+      "address": "0x37f03a12241E9FD3658ad6777d289c3fb8512Bc9",
+      "symbol": "ANT",
+      "id": "aragon",
+      "decimals": 18,
+      "precision": 2,
+      "chartColor": "#E7983D"
+    },
+    "0xccb0F4Cf5D3F97f4a55bb5f5cA321C3ED033f244": {
+      "address": "0xccb0F4Cf5D3F97f4a55bb5f5cA321C3ED033f244",
+      "symbol": "ZRX",
+      "id": "0x",
+      "decimals": 18,
+      "precision": 2,
+      "chartColor": "#536DFE"
+    }
+  },
+  "warnings": {},
+  "errors": {
+    "noBool": [
+      "0xEC4b4fFfb58B6c85D41025b07d93167B8192062E",
+      "0xe7d8E7742ca6906a14F7e46770Fdfe6B7014e62D"
+    ],
+    "transferFee": [
+      "0xbef02f140BFb37a75Ed6510e8D85C9E9e7e0ff1e"
+    ]
+  }
+}
