+++ src/helpers/lock.ts
@@ -1,9 +1,9 @@
 import { Lock } from '@bonustrack/lock/dist/lock.cjs';
-import config from '@/helpers/config';
 import injected from '@bonustrack/lock/connectors/injected';
 import portis from '@bonustrack/lock/connectors/portis';
 import walletconnect from '@bonustrack/lock/connectors/walletconnect';
 import walletlink from '@bonustrack/lock/connectors/walletlink';
+import config from '@/config';
 
 const connectors = { injected, portis, walletconnect, walletlink };
 const lock = new Lock();
