+++ src/helpers/ws.ts
@@ -1,5 +1,5 @@
 import { WebSocketProvider } from '@ethersproject/providers';
-import config from '@/helpers/config';
+import config from '@/config';
 
 const wsProvider = new WebSocketProvider(config.alchemyWsUrl);
